+++ yggdrasil_decision_forests/port/python/tools/local_copybara_export.sh
@@ -19,7 +19,7 @@
 #
 # Usage example:
 #   third_party/yggdrasil_decision_forests/port/python/tools/local_copybara_export.sh
-#   INTERRACTIVE=1 third_party/yggdrasil_decision_forests/port/python/tools/local_copybara_export.sh
+#   INTERACTIVE=1 third_party/yggdrasil_decision_forests/port/python/tools/local_copybara_export.sh
 #
 set -vex
 
@@ -43,7 +43,7 @@ function local_export () {
 
 build_and_test() {
   cd "${LOCAL_DIR}/yggdrasil_decision_forests/port/python"
-  sudo INTERRACTIVE=$INTERRACTIVE tools/release_linux.sh
+  sudo INTERACTIVE=$INTERACTIVE tools/release_linux.sh
 }
 
 local_export