+++ yggdrasil_decision_forests/port/python/ydf/model/generic_model.py
@@ -34,6 +34,7 @@ from ydf.model import analysis
 from ydf.model import model_metadata
 from ydf.model import optimizer_logs
 from ydf.model import template_cpp_export
+from ydf.utils import concurrency
 from ydf.utils import html
 from ydf.utils import log
 from yggdrasil_decision_forests.utils import model_analysis_pb2
@@ -380,7 +381,11 @@ Use `model.describe()` for more details
       return self._model.Serialize()
 
   def predict(
-      self, data: dataset.InputDataset, *, use_slow_engine=False
+      self,
+      data: dataset.InputDataset,
+      *,
+      use_slow_engine=False,
+      num_threads: Optional[int] = None,
   ) -> np.ndarray:
     """Returns the predictions of the model on the given dataset.
 
@@ -408,7 +413,12 @@ Use `model.describe()` for more details
         with the regular engines fail, e.g., models with a very large number of
         categorical conditions. It is only in these cases, that users should use
         the slow engine and report the issue to the YDF developers.
+      num_threads: Number of threads used to run the model.
     """
+
+    if num_threads is None:
+      num_threads = concurrency.determine_optimal_num_threads(training=False)
+
     with log.cc_log_context():
       # The data spec contains the label / weights /  ranking group / uplift
       # treatment column, but those are not required for making predictions.
@@ -417,7 +427,9 @@ Use `model.describe()` for more details
           data_spec=self._model.data_spec(),
           required_columns=self.input_feature_names(),
       )
-      result = self._model.Predict(ds._dataset, use_slow_engine)  # pylint: disable=protected-access
+      result = self._model.Predict(
+          ds._dataset, use_slow_engine, num_threads=num_threads
+      )  # pylint: disable=protected-access
     return result
 
   def evaluate(
@@ -431,6 +443,7 @@ Use `model.describe()` for more details
       bootstrapping: Union[bool, int] = False,
       evaluation_task: Optional[Task] = None,
       use_slow_engine: bool = False,
+      num_threads: Optional[int] = None,
   ) -> metric.Evaluation:
     """Evaluates the quality of a model on a dataset.
 
@@ -502,10 +515,14 @@ Use `model.describe()` for more details
         with the regular engines fail, e.g., models with a very large number of
         categorical conditions. It is only in these cases, that users should use
         the slow engine and report the issue to the YDF developers.
+      num_threads: Number of threads used to run the model.
 
     Returns:
       Model evaluation.
     """
+
+    if num_threads is None:
+      num_threads = concurrency.determine_optimal_num_threads(training=False)
 
     # Warning about deprecation of "evaluation_task"
     if evaluation_task is not None:
@@ -571,6 +588,7 @@ Use `model.describe()` for more details
       options_proto = metric_pb2.EvaluationOptions(
           bootstrapping_samples=bootstrapping_samples,
           task=task._to_proto_type(),  # pylint: disable=protected-access
+          num_threads=num_threads,
       )
 
       evaluation_proto = self._model.Evaluate(
@@ -580,6 +598,7 @@ Use `model.describe()` for more details
           label_col_idx=label_col_idx,
           group_col_idx=group_col_idx,
           use_slow_engine=use_slow_engine,
+          num_threads=num_threads,
       )  # pylint: disable=protected-access
     return metric.Evaluation(evaluation_proto)
 
@@ -637,7 +656,7 @@ Use `model.describe()` for more details
       partial_depepence_plot: bool = True,
       conditional_expectation_plot: bool = True,
       permutation_variable_importance_rounds: int = 1,
-      num_threads: int = 6,
+      num_threads: Optional[int] = None,
   ) -> analysis.Analysis:
     """Analyzes a model on a test dataset.
 
@@ -691,6 +710,9 @@ Use `model.describe()` for more details
     Returns:
       Model analysis.
     """
+
+    if num_threads is None:
+      num_threads = concurrency.determine_optimal_num_threads(training=False)
 
     with log.cc_log_context():
       ds = dataset.create_vertical_dataset(